requirements.txt
BeautifulSoup==3.2.1
cssselect==0.9.1
facepy==1.0.7
goose-extractor==1.0.25
jieba==0.37
lxml==3.4.4
nltk==3.1
Pillow==3.0.0
requests==2.8.1
six==1.10.0
wikiwords==0.5
